87 Split base.fixture.ts: extract diagnosis.util.ts and api.fixture.ts
