+++ playwright/typescript/tests/validation.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '@fixtures/base.fixture';
+import { test, expect } from '@fixtures/api.fixture';
 import { expectValidXhtml, expectValidCss } from '@utils/validation.util';
 import { PUBLIC_PAGE_CLASSES } from '@pages/public/index';
 import { AUTHENTICATED_PAGE_CLASSES } from '@pages/authenticated/index';
